fix(ci): split task ws helpers and unpoison zero-count
Keep max-lines on tasks.ts under the frontend lint limit. On
zero-count persist failure, restore prior queuedCount and only
suppress verified gone-task FK errors so retries still work.
